moved PlotPTMaps
