+++ mtpy/modeling/ModEM/__init__.py
@@ -9,9 +9,10 @@ from .convariance import Covariance
 from .config import ModEMConfig
 from .model_manipulator import ModelManipulator
 from .plot_response import PlotResponse
+from .plot_pt_maps import PlotPTMaps
 
 __all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual',
            'ControlInv', 'ControlFwd', 'Covariance', 'ModEMConfig', 'ModelManipulator',
-           'PlotResponse']
+           'PlotResponse', 'PlotPTMaps']
